pyproject.toml
[tool.poetry]
name = "knotty"
version = "0.1.0"
description = "A package repository"
authors = ["slowlime"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
sqlalchemy = "^2.0.9"
fastapi = "^0.95.0"
passlib = {extras = ["argon2"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
toml = "^0.10.2"

[tool.poetry.group.dev.dependencies]
mypy = "^1.2.0"
black = "^23.3.0"
ruff = "^0.0.261"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
